separated shape and norm hist production

+++ Analysis/HistProducerSample.py
@@ -15,14 +15,10 @@ if __name__ == "__main__":
     import argparse
     import yaml
     parser = argparse.ArgumentParser()
-    parser.add_argument('--inputDir', required=True, type=str)
     parser.add_argument('--histDir', required=True, type=str)
     parser.add_argument('--test', required=False, type=bool, default=False)
-    parser.add_argument('--deepTauVersion', required=False, type=str, default='v2p1')
+    parser.add_argument('--remove-files', required=False, type=bool, default=False)
     parser.add_argument('--dataset', required=True, type=str)
-    parser.add_argument('--compute_unc_variations', type=bool, default=False)
-    parser.add_argument('--compute_rel_weights', type=bool, default=False)
-    parser.add_argument('--histConfig', required=True, type=str)
     parser.add_argument('--sampleConfig', required=True, type=str)
     args = parser.parse_args()
 
@@ -32,10 +28,6 @@ if __name__ == "__main__":
 
     if not os.path.isdir(args.histDir):
         os.makedirs(args.histDir)
-    hist_cfg_dict = {}
-    with open(args.histConfig, 'r') as f:
-        hist_cfg_dict = yaml.safe_load(f)
-    vars_to_plot = list(hist_cfg_dict.keys())
 
     # 1 list files :
 
@@ -46,45 +38,36 @@ if __name__ == "__main__":
     if(args.dataset=='data'): sample_type = 'data'
     else: sample_type = sample_cfg_dict[args.dataset]['sampleType']
 
-    all_files = os.listdir(os.path.join(args.inputDir, args.dataset))
-    for single_file in all_files:
-        inputFile = os.path.join(args.inputDir, args.dataset, single_file)
-        if not os.path.isdir(tmpDir):
-            os.makedirs(tmpDir)
-        print(inputFile)
-        histProducerFile = os.path.join(os.environ['ANALYSIS_PATH'], "Analysis/HistProducerFile.py")
-        cmd_list = ['python3', histProducerFile, '--inFile', inputFile, '--outDir', tmpDir,
-                    '--dataset', args.dataset, '--histConfig', args.histConfig]
-        if args.compute_unc_variations:
-            cmd_list.extend(['--compute_unc_variations', 'True'])
-        if args.compute_rel_weights:
-            cmd_list.extend(['--compute_rel_weights', 'True'])
-        print(cmd_list)
-        if args.test:
-            cmd_list.extend(['--test','True'])
-        sh_call(cmd_list,verbose=1)
+    all_dir_vars = os.listdir(tmpDir)
 
-    for var in vars_to_plot:
-        if var not in samples_files.keys():
-            samples_files[var] = {}
-        hist_produced_dir = os.path.join(tmpDir, var)
-        all_out_files = []
-        for hist in os.listdir(hist_produced_dir):
-            if args.dataset in hist:
-                all_out_files.append(os.path.join(hist_produced_dir, hist))
-        hist_out_dir = os.path.join(args.histDir, var)
-        if not os.path.isdir(hist_out_dir):
-            os.makedirs(hist_out_dir)
-        outFileName = f'{hist_out_dir}/{args.dataset}.root'
+    for dir_var in all_dir_vars:
+        txtFile  = os.path.join(tmpDir, dir_var, "filesToHAdd.txt")
+        input_files_to_write = []
+        input_files_to_read = []
+        for root, dirs, files in os.walk(os.path.join(tmpDir, dir_var)):
+            for file in files:
+                if file.endswith('.root') and not file.startswith('.'):
+                    if os.path.join(root, file) not in input_files_to_read:
+                        input_files_to_write.append(os.path.join(root, file))
+        with open(txtFile, 'w') as inputFileTxt:
+            for input_line in input_files_to_write:
+                inputFileTxt.write(input_line+'\n')
+        with open(txtFile, 'r') as inputtxtFile:
+            input_files_to_read = inputtxtFile.read().splitlines()
+            if len(input_files_to_read) == 0:
+                raise RuntimeError(f"no input files found for {tmpDir}, {dir_var}")
+
+        outFileName = f'{args.histDir}/{args.dataset}.root'
         hadd_str = f'hadd -f209 -j -O {outFileName} '
-        hadd_str += ' '.join(f for f in all_out_files)
-        if len(all_out_files) > 1:
+        hadd_str += ' '.join(f for f in input_files_to_read)
+        if len(input_files_to_read) > 1:
             sh_call([hadd_str], True)
-            if os.path.exists(outFileName):
-                for histFile in all_out_files:
-                    if args.test : print(histFile)
-                    if histFile == outFileName: continue
-                    os.remove(histFile)
         else:
-            shutil.move(all_out_files[0],outFileName)
-    shutil.rmtree(os.path.join(args.histDir, args.dataset))
+            shutil.copy(input_files_to_read[0],outFileName)
+        if os.path.exists(outFileName) and args.remove_files:
+            for histFile in input_files_to_read:
+                if args.test : print(histFile)
+                if histFile == outFileName: continue
+                os.remove(histFile)
+    if args.remove_files:
+        shutil.rmtree(tmpDir)
